requirements.txt
python-telegram-bot>=22.6
openai>=2.0
python-dotenv>=1.0.0
